Set destination folder to be root directory of ADS

+++ AxonDeepSeg/models/download_model.py
@@ -5,6 +5,7 @@ import argparse
 from argparse import RawTextHelpFormatter
 
 def download_model(destination = None):
+    root_directory = False             #Flag to check if the destination folder is the root directory 
     if destination is None:
         sem_destination = Path("AxonDeepSeg/models/default_SEM_model")
         tem_destination = Path("AxonDeepSeg/models/default_TEM_model")
@@ -12,9 +13,19 @@ def download_model(destination = None):
         destination = convert_path(destination)
         sem_destination = destination / "default_SEM_model"
         tem_destination = destination / "default_TEM_model"
-
+    print("SEM destination is", sem_destination)
+    print("TEM destination is", tem_destination)
     url_TEM_model = "https://osf.io/2hcfv/?action=download&version=5"  # URL of TEM model hosted on OSF storage with the appropriate versioning on OSF
     url_SEM_model = "https://osf.io/sv7u2/?action=download&version=5"  # URL of SEM model hosted on OSF storage with the appropriate versioning on OSF
+
+    if destination is not None and destination.name =="axondeepseg":     #Deletes existing models if present in the root directory 
+        root_directory = True
+        if sem_destination.exists():
+            print('SEM model folder already existed - deleting old one.')
+            shutil.rmtree(str(sem_destination))
+        if tem_destination.exists():
+            print('TEM model folder already existed - deleting old one.')
+            shutil.rmtree(str(tem_destination))
 
     if (
         not download_data(url_TEM_model) and not download_data(url_SEM_model)
@@ -25,16 +36,19 @@ def download_model(destination = None):
             "ERROR: Data was not succesfully downloaded and unzipped- please check your link and filename and try again."
         )
 
-    if sem_destination.exists():
-        print('SEM model folder already existed - deleting old one.')
-        shutil.rmtree(str(sem_destination))
-    if tem_destination.exists():
-        print('TEM model folder already existed - deleting old one.')
-        shutil.rmtree(str(tem_destination))
+    if not root_directory:           # if destination folder is other than root directory
+        if sem_destination.exists():
+            print('SEM model folder already existed - deleting old one.')
+            shutil.rmtree(str(sem_destination))
+        if tem_destination.exists():
+            print('TEM model folder already existed - deleting old one.')
+            shutil.rmtree(str(tem_destination))
 
-    shutil.move("default_SEM_model", str(sem_destination))
-    shutil.move("default_TEM_model", str(tem_destination))
+        shutil.move("default_SEM_model", str(sem_destination))
+        shutil.move("default_TEM_model", str(tem_destination))
 
+    print("Models are stored in the directory: ", sem_destination.parent)
+    
 
 def main(argv=None):
